requirements.txt
# authlib==0.14.2
# authutils==4.0.0
# boto==2.36.0
# cryptography==2.8
# dictionaryutils==3.1.0
# envelopes==0.4
# Flask==1.1.1
# flask-cors==3.0.3
# Flask-SQLAlchemy-Session==1.1
# fuzzywuzzy==0.6.1
# gen3authz==0.2.1
# graphene==2.0.1
# jsonschema==3.2
# lxml==4.5.2
# pbr==2.0.0
# psycopg2==2.8.4
# python-keystoneclient==1.8.1
# PyYAML==5.1
# requests==2.22.0
# setuptools==40.3.0
# six==1.12.0
# simplejson==3.8.1
# sqlalchemy==1.3.5
# Werkzeug==0.16.0
# xmltodict==0.9.2
# more-itertools==5.0.0
# cdis_oauth2client==1.0.0
# cdispyutils==1.0.4
# datamodelutils==1.0.0
# # required for gen3datamodel, not required for sheepdog
# psqlgraph==3.0.0
# cdiserrors==0.1.2
# cdislogging==1.0.0
# git+https://git@github.com/NCI-GDC/cdisutils.git@f54e393c89939b2200dfae45c6235cbe2bae1206#egg=cdisutils
# gen3dictionary==2.0.1
# gen3datamodel==3.0.2
# git+https://git@github.com/uc-cdis/indexclient.git@2.0.0#egg=indexclient
# git+https://git@github.com/uc-cdis/storage-client.git@1.0.0#egg=storageclient
# scipy==1.3.1
# lifelines==0.25.2
pycryptodome==3.9.8
